Modifiers working both on 32bits and 64bits to display uintptr_t
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/simgrid/simgrid/trunk@7421 48e7efb5-ca39-0410-a469-dd3cf9ba447f

+++ src/simix/smx_network.c
@@ -389,7 +389,7 @@ void SIMIX_message_sizes_output(const char *filename) {
   uintptr_t key,data;
   xbt_dict_cursor_t cursor;
   xbt_dict_foreach(msg_sizes,cursor,key,data) {
-    fprintf(out,"%ld %ld\n",key,data);
+    fprintf(out,"%ld %ld\n",(long unsigned)key,(long unsigned)data);
   }
   fclose(out);
 }
